draft init module
