+++ src/utils/goals.ts
